requirements.txt
matplotlib==3.10.7
pandas==2.3.3
pyarrow==21.0.0
pygame==2.6.1
python-chess==1.999
# torch and torchvision must be installed separately
